# src/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ResJson_CheckInRes_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"result": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"payload": {
						"$ref": "#/components/schemas/CheckInRes"
					}
				},
				"required": [
					"status",
					"code",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorJson": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"result": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"code",
					"result",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResJson_null_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"result": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"payload": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"status",
					"code",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UserStatusResponse": {
				"properties": {
					"user": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/UserStatus"
							},
							{
								"$ref": "#/components/schemas/UserStatusJSON"
							}
						]
					},
					"isAdmin": {
						"type": "boolean"
					}
				},
				"required": [
					"user",
					"isAdmin"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResJson_UserStatusResponse_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"result": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"payload": {
						"$ref": "#/components/schemas/UserStatusResponse"
					}
				},
				"required": [
					"status",
					"code",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"CookieAuth": {
				"type": "apiKey",
				"name": "w_auth_local",
				"in": "cookie"
			},
			"OAuth42": {
				"type": "oauth2",
				"flows": {
					"clientCredentials": {
						"type": "oauth2",
						"authorizationUrl": "http://localhost:3000/user/login",
						"tokenUrl": "http://localhost:3000/user/login/token",
						"refreshUrl": "http://localhost:3000/user/token/refresh",
						"scopes": []
					}
				}
			}
		}
	},
	"info": {
		"title": "42checkin",
		"version": "2.2.7",
		"description": "Cluster checkin service for 42 Seoul cadets.",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "@ryukim, @yurlee, @ohjongin"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/v1/user/checkIn/{cardId}": {
			"post": {
				"operationId": "CheckIn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResJson_CheckInRes_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 200,
											"result": true,
											"code": 2000,
											"payload": {
												"card_no": 10,
												"state": "checkIn",
												"prev_state": "checkOut",
												"notice": false
											}
										}
									},
									"Example 2": {
										"value": {
											"status": 200,
											"result": false,
											"code": 4090,
											"message": "이미 체크인 상태입니다",
											"payload": {
												"card_no": 10,
												"state": "checkIn",
												"prev_state": "checkIn",
												"notice": false
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 401,
											"code": 4010,
											"message": "사용자 정보 없음",
											"result": false
										}
									}
								}
							}
						}
					},
					"406": {
						"description": "Not Acceptable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 406,
											"code": 4060,
											"message": "체크인 가능 시간이 아닙니다.",
											"result": false
										}
									}
								}
							}
						}
					},
					"409": {
						"description": "Not Acceptable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 409,
											"code": 4090,
											"message": "이미 사용 중인 카드입니다. cardOwner: cadet",
											"result": false
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 500,
											"code": 5000,
											"message": "Internal Server Error",
											"result": false
										}
									}
								}
							}
						}
					}
				},
				"description": "유저 체크인",
				"tags": [
					"User"
				],
				"security": [
					{
						"CookieAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "cardId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/user/checkOut": {
			"post": {
				"operationId": "CheckOut",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResJson_null_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 200,
											"code": 2000,
											"result": true,
											"message": "checkout"
										}
									},
									"Example 2": {
										"value": {
											"status": 200,
											"code": 4090,
											"result": false,
											"message": "이미 체크아웃 하셨습니다."
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 401,
											"code": 4010,
											"result": false,
											"message": "사용자 정보 없음"
										}
									}
								}
							}
						}
					},
					"406": {
						"description": "Not Acceptable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 406,
											"code": 4060,
											"message": "체크아웃 가능 시간이 아닙니다.",
											"result": false
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 500,
											"code": 5000,
											"message": "Internal Server Error",
											"result": false
										}
									}
								}
							}
						}
					}
				},
				"description": "유저 체크아웃",
				"tags": [
					"User"
				],
				"security": [
					{
						"CookieAuth": []
					}
				],
				"parameters": []
			}
		},
		"/v1/user/status": {
			"get": {
				"operationId": "Status",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResJson_UserStatusResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 200,
											"code": 200,
											"result": true,
											"payload": {
												"user": {
													"login": "chaepark",
													"card": 25,
													"state": "checkIn",
													"log_id": 5,
													"checkin_at": "2022-02-02T07:59:24.000Z",
													"checkout_at": null,
													"profile_image_url": "https://cdn.intra.42.fr/users/chaepark.jpg"
												},
												"isAdmin": false
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 401,
											"code": 4010,
											"result": false,
											"message": "사용자 정보 없음"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 500,
											"code": 5000,
											"message": "Internal Server Error",
											"result": false
										}
									}
								}
							}
						}
					}
				},
				"description": "유저 상태조회",
				"tags": [
					"User"
				],
				"security": [
					{
						"CookieAuth": []
					}
				],
				"parameters": []
			}
		},
		"/v1/user/usage": {
			"get": {
				"operationId": "UserUsageList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResJson_UsageAttributes-Array_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 200,
											"code": 2000,
											"result": true,
											"payload": [
												{
													"_id": 1,
													"login": "chaepark",
													"checkin_at": "2022-02-02T07:29:20.000Z",
													"checkout_at": "2022-02-02T07:39:41.000Z",
													"duration": 622,
													"actor": "chaepark",
													"deleted_at": null,
													"updated_at": null,
													"created_at": "2022-02-02T07:39:41.000Z"
												},
												{
													"_id": 2,
													"login": "chaepark",
													"checkin_at": "2022-02-02T07:39:59.000Z",
													"checkout_at": "2022-02-02T07:58:15.000Z",
													"duration": 1097,
													"actor": "chaepark",
													"deleted_at": null,
													"updated_at": null,
													"created_at": "2022-02-02T07:58:15.000Z"
												}
											]
										}
									},
									"Example 2": {
										"value": {
											"status": 200,
											"code": 4040,
											"result": false,
											"message": "주어진 기간 내 유저 사용 정보가 없습니다.",
											"payload": []
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								},
								"examples": {
									"Example 1": {
										"value": "Unauthorized"
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 500,
											"code": 5000,
											"message": "Internal Server Error",
											"result": false
										}
									}
								}
							}
						}
					}
				},
				"description": "기간 내 유저 사용 정보 조회",
				"tags": [
					"User"
				],
				"security": [
					{
						"CookieAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "from",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "2022-01-31"
					},
					{
						"in": "query",
						"name": "to",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "2022-02-01"
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "localhost"
		},
		{
			"url": "https://api.checkin.dev.42seoul.io",
			"description": "development"
		}
	]
}